26.06.22 refactor: 整理代码结构并修复多处细节问题
本次提交包含多项改进：
1.  重构骑士美德计数逻辑，提取为独立工具类方法
2.  统一分支统计说明的缩进常量，替换硬编码空格
3.  添加新的坚韧二阶美德类
4.  优化Hediff和美德处理器的回调通知逻辑
5.  重构骑士交谈互动逻辑，拆分并复用代码
6.  修复Hediff模板的空值检查和缓存问题
7.  调整居民骑士初始化逻辑，移入PostAdded方法
8.  清理多余空行和简化循环代码

26.06.20
feat: 新增骑士美德系统与相关功能优化

1.  新增ITickInterval接口与骑士美德Hediff
2.  重构骑士美德生成与生命周期管理
3.  优化荣誉圣骑士Hediff逻辑
4.  为骑士团与常驻骑士逻辑添加异常捕获
5.  修复常驻骑士禁用状态判断

26.06.18
refactor(KnightVirtue): 重构美德词条系统并修复等级限制问题

1. 新增KnightVirtueTraitGroups统一管理词条组，替换原有的分级词条字段
2. 调整美德等级下限从0改为1，统一初始化与存档默认值
3. 简化属性缓存调用逻辑，移除冗余空判断
4. 重构词条属性加成应用逻辑，提取公共方法优化代码结构
5. 简化配置校验逻辑，移除固定3选1的硬编码限制

refactor: 统一重构交互与物品生成相关代码

1.  将多处OAFrame_MiscUtility.TryGenerateThing替换为OAFrame_ThingUtility.GenerateThingListSplitByStack
2.  重构交互工作类的方法命名，统一使用ApplyEffect替代ApplyInteraction
3.  简化回调方法为表达式体并添加异常捕获日志
4.  优化采矿探索交互的常量使用与代码格式

+++ Utility/WindowUtility.cs
@@ -382,7 +382,7 @@ public static class OARO_WindowUtility
             Color color = material.shader.SupportsMaskTex() ? GUI.color : new Color(GUI.color.r * 0.5f, GUI.color.g * 0.5f, GUI.color.b * 0.5f, GUI.color.a);
             Rect screenRect = default;
             Rect sorceRect = default;
-            float imageAspect = (float)texture.width / (float)texture.height;
+            float imageAspect = texture.width / (float)texture.height;
             CalculateScaledTextureRects(rect, scaleMode, imageAspect, ref screenRect, ref sorceRect);
             Graphics.DrawTexture(screenRect, texture, sorceRect, 0, 0, 0, 0, color, material);
         }
